Clean up compile warnings (which exposed a few errors).  Clearly I need to use Make to build.

+++ Turret/arduino/turret/turretController.h
@@ -120,7 +120,7 @@ private:
     const uint32_t k_safeStateMinDt = 500000;
     const uint32_t k_selfRightTriggerDt = 3000000;
 
-    const uint8_t k_throwIntensityToSwingTime[9] = { 1000000, 900000, 800000, 700000, 600000, 500000, 400000, 300000, 200000 };
+    const uint32_t k_throwIntensityToSwingTime[9] = { 1000000, 900000, 800000, 700000, 600000, 500000, 400000, 300000, 200000 };
